Allow gcc-arm-embedded to be anywhere on system PATH (#1)
Scons uses its own PATH by default, so we have to explicitly use the system PATH (os.environ['PATH']). This allows us to not have to specify the location of the gcc-arm-embedded tools (which is different between backpack and packer boxes anyways), making it build without modification on my system.