[FWXV-13] Task testing helper library (#28)
It's a helper library for creating task tests, which are tests that run in the context of a highest-priority task.

We define an MS_TEST preprocessor symbol when in test mode now, and we use this in FreeRTOSConfig.h to add a new highest priority so that the test task is always the highest priority. Also, users can redefine the max priority using the NUM_FREERTOS_PRIORITIES preprocessor symbol.