current sense soc work
It worked at some point, but now doesn't. Current broken point is that
configuration doesn't seem to (fully?) apply. VEmpty seems to be set
correctly, but the design capacity is reporting back 3000 (30000mah with
our rsense) which is the default.

I'm not sure why this is happening, it definitely worked at some point
today, no idea why it isn't now.